add ( catalog_screen ) : add InputPasswordFieldUserProfile , InputTextFileUserProfile
