+++ lib/presentation/screens/catalog/widgets/inputs_user_v2/input_text_v2.dart
@@ -1,0 +1,71 @@
+import 'package:finniu/presentation/providers/settings_provider.dart';
+import 'package:flutter/material.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+
+class InputTextFileUserProfile extends ConsumerWidget {
+  const InputTextFileUserProfile({
+    super.key,
+    required this.controller,
+    required this.hintText,
+    required this.validator,
+    this.isNumeric = false,
+  });
+
+  final TextEditingController controller;
+  final String? Function(String?)? validator;
+  final String hintText;
+  final bool isNumeric;
+
+  final int hintDark = 0xFF989898;
+  final int hintLight = 0xFF989898;
+  final int fillDark = 0xFF222222;
+  final int fillLight = 0xFFF7F7F7;
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final isDarkMode = ref.watch(settingsNotifierProvider).isDarkMode;
+
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        TextFormField(
+          controller: controller,
+          decoration: InputDecoration(
+            hintStyle: TextStyle(
+              fontSize: 12,
+              color: isDarkMode ? Color(hintDark) : Color(hintLight),
+              fontWeight: FontWeight.w400,
+              fontFamily: "Poppins",
+            ),
+            hintText: hintText,
+            fillColor: isDarkMode ? Color(fillDark) : Color(fillLight),
+            filled: true,
+            border: const OutlineInputBorder(
+              borderRadius: BorderRadius.all(Radius.circular(25.0)),
+              borderSide: BorderSide.none,
+            ),
+            enabledBorder: const OutlineInputBorder(
+              borderRadius: BorderRadius.all(Radius.circular(25.0)),
+              borderSide: BorderSide.none,
+            ),
+            focusedBorder: const OutlineInputBorder(
+              borderRadius: BorderRadius.all(Radius.circular(25.0)),
+              borderSide: BorderSide.none,
+            ),
+            errorBorder: const OutlineInputBorder(
+              borderRadius: BorderRadius.all(Radius.circular(25.0)),
+              borderSide: BorderSide.none,
+            ),
+            focusedErrorBorder: const OutlineInputBorder(
+              borderRadius: BorderRadius.all(Radius.circular(25.0)),
+              borderSide: BorderSide.none,
+            ),
+          ),
+          validator: validator,
+          keyboardType: isNumeric ? TextInputType.number : TextInputType.text,
+        ),
+        const SizedBox(height: 15),
+      ],
+    );
+  }
+}
